Use search func inside component
