--- src/components/widgets/search-widget/search-widget-styles.module.scss
@@ -1,52 +1,0 @@
-@import 'styles/settings';
-@import 'styles/ui.module';
-@import 'styles/typography-extends';
-
-.searchButton {
-
-  align-items: center;
-  color: $white;
-  display: flex;
-  font-weight: bold;
-  height: $widget-button-size;
-  left: $site-gutter;
-  padding: 0;
-  position: absolute;
-  top: $site-gutter;
-
-  // .iconWrapper{
-  //   align-items: center;
-  //   background-color: $elephant;
-  //   border-radius: 2px;
-  //   display: flex;
-  //   height: $widget-button-size;
-  //   width: $widget-button-size;
-  //   justify-content: center;
-  //   margin-right: 10px;
-  //   padding: 0;
-  // }
-
-  .searchCardTitle {
-    @extend %display1;
-    margin-top: 2px;
-    padding-right: 20px;
-  }
-
-  .searchSubtitle {
-    @extend %title;
-    margin-top: 2px;
-    padding-right: 20px;
-  }
-
-  // svg {
-  //   fill: $white;
-  //   height: 16px;
-  //   width: 16px;
-  // }
-}
-
-@media print {
-  .searchButton {
-    display: none;
-  }
-}
