+++ src/styles/override/esri-search-widget.scss
@@ -7,7 +7,6 @@
   height: 345px;
   bottom: 55px;
   top: auto;
-  border-radius: $border-radius-mobile;
   position: fixed;
   left: 0;
   padding: 30px 20px;
