+++ src/styles/settings.scss
@@ -128,7 +128,7 @@ $letter-spacing: 1.5px;
 $line-height: 14px;
 
 // Sidebar
-$sidebar-top-margin: 60px;
+$sidebar-top-margin: 25px;
 $local-sidebar-top-margin: 20px;
 $high-sidebar-top-margin: 268px;
 $sidebar-header-height: 100px;
